requirements.txt
asgiref==3.2.7
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dj-database-url==0.5.0
Django==3.0.5
django-cors-headers==3.4.0
django-heroku==0.3.1
djangorestframework==3.11.0
h11==0.9.0
heroku==0.1.4
idna==2.10
psycopg2==2.8.5
python-dateutil==1.5
pytz==2019.3
requests==2.24.0
sqlparse==0.3.1
syncasync==20180812
urllib3==1.25.9
uvicorn==0.11.5
websockets==8.1
whitenoise==5.1.0
